שיפורי סינטקס - בעיקרס בסרוויס המיילים, הסרת תווי בריחה מיותרים מרגקס'ס
